axiom/ai_layer/services/requirements-langgraph.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
neo4j>=5.22.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.7.0